Update README.md
Quiz
Developed a quiz that provides an opportunity to test the knowledge of users. The quiz checks the answers and calculates the points scored. If the answer is correct, a message about correctness is displayed, and with an incorrect answer - a message about infidelity. At the end of the quiz, the user receives information about their final score.